adjust to changed dirEntries-function: throws now if dir does not exist
git-svn-id: http://svn.dsource.org/projects/visuald/trunk@176 a09178c5-0f50-412e-aa4a-0aec1a98dccb

+++ visuald/build.d
@@ -446,9 +446,10 @@ else
 				{
 					string dir = dirname(file);
 					string pattern = basename(file);
-					foreach(string f; dirEntries(dir, SpanMode.shallow))
-						if(fnmatch(f, pattern))
-							std.file.remove(f);
+					if(isExistingDir(dir))
+						foreach(string f; dirEntries(dir, SpanMode.shallow))
+							if(fnmatch(f, pattern))
+								std.file.remove(f);
 				}
 				else if(std.file.exists(file))
 					std.file.remove(file);
